+++ src/publisher.rs
@@ -1,0 +1,6 @@
+use crate::share::Id;
+
+pub struct Publisher{
+    name: Option<String>,
+    books: Option<Vec<Id>>,
+}
