+++ src/series.rs
@@ -1,0 +1,4 @@
+pub struct Series{
+    title: Option<String>,
+    book_ids: Vec<usize>,
+}
