+++ src/share.rs
@@ -1,0 +1,2 @@
+pub type Id = usize;
+pub type Date = ();
